requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
distlib==0.3.8
fake-useragent==1.5.1
Faker==25.8.0
filelock==3.15.4
h11==0.14.0
idna==3.6
lxml==5.1.0
MouseInfo==0.1.3
outcome==1.3.0.post0
pillow==10.2.0
platformdirs==4.2.2
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
requests==2.31.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
virtualenv==20.26.3
wsproto==1.2.0
